add local linux build for protobuf add QNX support for up-core-api
