+++ up-core-api/developer/conanfile.py
@@ -29,7 +29,11 @@ class upCoreApiRecipe(ConanFile):
             "fork": "eclipse-uprotocol/up-spec",
             "commitish": "main"}
 
-    requires = "protobuf/[~3.21]"
+    def requirements(self):
+        self.requires("protobuf/3.21.12")
+
+    def build_requirements(self):
+        self.tool_requires("protobuf/3.21.12")
 
     def init(self):
         self.fork = self.options.get_safe("fork", "eclipse-uprotocol/up-spec")
@@ -58,16 +62,31 @@ class upCoreApiRecipe(ConanFile):
         deps = CMakeDeps(self)
         deps.generate()
         tc = CMakeToolchain(self)
+        if self.settings.os == "Neutrino":
+            tc.cache_variables["CMAKE_SYSTEM_NAME"] = "QNX"
+            tc.cache_variables["CMAKE_C_COMPILER"] = "qcc"
+            tc.cache_variables["CMAKE_CXX_COMPILER"] = "q++"
+            tc.cache_variables["CMAKE_CXX_STANDARD"] = 14
+            if   self.settings.arch == "armv8": #aarch64le
+                tc.cache_variables["CMAKE_SYSTEM_PROCESSOR"] = "aarch64le"
+                tc.cache_variables["CMAKE_CXX_COMPILER_TARGET"] = "gcc_ntoaarch64le"
+                tc.cache_variables["CMAKE_C_COMPILER_TARGET"] = "gcc_ntoaarch64le"
+            elif self.settings.arch == "x86_64": #x86_64
+                tc.cache_variables["CMAKE_SYSTEM_PROCESSOR"] = "x86_64"
+                tc.cache_variables["CMAKE_CXX_COMPILER_TARGET"] = "gcc_ntox86_64"
+                tc.cache_variables["CMAKE_C_COMPILER_TARGET"] = "gcc_ntox86_64"
+            tc.cache_variables["CMAKE_CXX_FLAGS_INIT"] = "-D_XOPEN_SOURCE=700 -D_QNX_SOURCE"
+            tc.cache_variables["CMAKE_C_FLAGS_INIT"] = "-D_XOPEN_SOURCE=700 -D_QNX_SOURCE"
         tc.generate()
 
     def build(self):
         cmake = CMake(self)
         cmake.configure()
-        cmake.build()
+        cmake.build(cli_args=["--verbose"])
 
     def package(self):
         cmake = CMake(self)
-        cmake.install()
+        cmake.install(cli_args=["--verbose"])
 
     def package_info(self):
         self.cpp_info.libs = ["up-core-api"]
